KGFC/fact_checker/requirements.txt
google-generativeai>=0.3.0
sparqlwrapper>=2.0.0
sentence-transformers>=2.2.0
spacy>=3.7.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
tqdm>=4.65.0